requirements.txt
selenium
pyautogui
webdriver-manager
pywin32
Pillow